Done with bulk media

+++ app/Http/Controllers/AdminMediasController.php
@@ -15,14 +15,11 @@ class AdminMediasController extends Controller
     {
 
         $photos = Photo::all();
-        $photos  = Photo::paginate(10);
+        $photos = Photo::paginate(10);
 
 
-
-       return view('admin.medias.index', compact('photos'));
+        return view('admin.medias.index', compact('photos'));
     }
-
-
 
 
     public function create()
@@ -34,21 +31,50 @@ class AdminMediasController extends Controller
 
     public function store(Request $request)
     {
+
         $file = $request->file('file');
         $name = time() . $file->getClientOriginalName();
 
         $file->move('images', $name);
-        Photo::create(['path'=>$name]);
+        Photo::create(['path' => $name]);
 
 
     }
 
-    public  function destroy($id)
+    public function destroy($id)
     {
         $photo = Photo::findorfail($id);
-        unlink(public_path(). $photo->path);
-      $photo->delete() ;
-      Session::flash('deleted_photo', 'The photo has been deleted');
-        return redirect('admin/medias');
+        unlink(public_path() . $photo->path);
+        $photo->delete();
+        Session::flash('deleted_photo', 'The photo has been deleted');
+//        return redirect('admin/medias');
     }
+
+
+    public function deleteMedia(Request $request)
+    {
+
+        // if there is a button with the name delete_single
+//        if (isset($request->delete_single)) {
+//            $photoId = array_search('Delete', $request->delete_single);
+//            $this->destroy($photoId);
+//            return redirect()->back();
+//        }
+
+        if($request->delete_all && !empty($request->checkBoxArray)){
+
+           $photos = Photo::findorfail($request->checkBoxArray);
+            foreach($photos as $photo)
+               {
+                          $photo->delete();
+               }
+            return redirect()->back();
+
+
+        }
+        else return redirect()->back();
+
+    }
+
 }
+
